requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-chroma
pdfplumber
python-dotenv
chromadb